separate astar
