+++ qdpf.cpp
@@ -55,12 +55,15 @@ IDirectedGraph<int> *AStarPathFinder::GetGateGraph() { return pImpl->GetGateGrap
 std::size_t AStarPathFinder::NodePathSize() const { return pImpl->NodePath().size(); }
 void AStarPathFinder::Reset(int x1, int y1, int x2, int y2) { pImpl->Reset(x1, y1, x2, y2); }
 int AStarPathFinder::ComputeNodeRoutes() { return pImpl->ComputeNodeRoutes(); }
+
 void AStarPathFinder::VisitComputedNodeRoutes(NodeVisitor &visitor) const {
   for (auto [node, cost] : pImpl->NodePath()) visitor(node->x1, node->y1, node->x2, node->y2);
 }
+
 int AStarPathFinder::ComputeGateRoutes(CellCollector &collector, bool useNodePath) {
   return pImpl->ComputeGateRoutes(collector, useNodePath);
 }
+
 void AStarPathFinder::ComputePathToNextRouteCell(int x1, int y1, int x2, int y2,
                                                  CellCollector &collector) const {
   pImpl->ComputePathToNextRouteCell(x1, y1, x2, y2, collector);
